[WebUI]: Improve overriding approach to be more generic
