+++ browser/resources/downloads/brave_overrides/index.ts
@@ -3,6 +3,4 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
 // You can obtain one at https://mozilla.org/MPL/2.0/.
 
-import 'chrome://resources/brave/br_elements/br_toolbar/br_toolbar.js'
-
 import './item.js'
