+++ browser/resources/settings/brave_overrides/index.ts
@@ -16,12 +16,10 @@ import './autofill_page.js'
 import './appearance_page.js'
 import './basic_page.js'
 import './clear_browsing_data_dialog.js'
-import './cr_icon.js'
 import './add_site_dialog.js'
 import './cookies_page.js'
 import './default_browser_page.js'
 import './import_data_dialog.js'
-import './iron_icon.js'
 import './page_visibility.js'
 import './passwords_section.js'
 import './payments_section.js'
