+++ browser/resources/settings/sources.gni
@@ -74,11 +74,9 @@ brave_settings_non_web_component_files = [
   "brave_overrides/clear_browsing_data_dialog.ts",
   "brave_overrides/config.ts",
   "brave_overrides/cookies_page.ts",
-  "brave_overrides/cr_icon.ts",
   "brave_overrides/default_browser_page.ts",
   "brave_overrides/import_data_dialog.ts",
   "brave_overrides/index.ts",
-  "brave_overrides/iron_icon.ts",
   "brave_overrides/theme_color_picker.ts",
   "brave_overrides/page_visibility.ts",
   "brave_overrides/passwords_section.ts",
