+++ browser/ui/webui/ethereum_remote_client/ethereum_remote_client_page.html
@@ -11,7 +11,7 @@
   <script src="chrome://resources/js/load_time_data_deprecated.js"></script>
   <script src="/strings.js"></script>
   <script src="/ethereum_remote_client_page.bundle.js"></script>
-  <script type="module" src="chrome://resources/brave/br_elements/br_toolbar/br_toolbar.html.js"></script>
+  <script type="module" src="chrome://resources/cr_elements/cr_toolbar/cr_toolbar.html.js"></script>
   <style>
     #root {
       display: flex;
