+++ chromium_src/check_chromium_src_config.json5
@@ -115,12 +115,6 @@
         "ui/color/color_id.h": [
             "BRAVE_MATERIAL_MISSING_COLOR_IDS"
         ],
-        "ui/resources/grit/webui_resources.h": [
-            "IDR_WEBUI_CR_ELEMENTS_CR_BUTTON_CR_BUTTON_HTML_JS",
-            "IDR_WEBUI_CR_ELEMENTS_CR_BUTTON_CR_BUTTON_JS",
-            "IDR_WEBUI_CR_ELEMENTS_CR_TOGGLE_CR_TOGGLE_HTML_JS",
-            "IDR_WEBUI_CR_ELEMENTS_CR_TOGGLE_CR_TOGGLE_JS",
-        ],
         "v8/src/builtins/builtins-utils.h": [
             "BUILTIN",
         ],
