--- chromium_src/ui/resources/grit/webui_resources_map.cc
@@ -1,10 +1,0 @@
-// Copyright (c) 2024 The Brave Authors. All rights reserved.
-// This Source Code Form is subject to the terms of the Mozilla Public
-// License, v. 2.0. If a copy of the MPL was not distributed with this file,
-// You can obtain one at https://mozilla.org/MPL/2.0/.
-
-// webui_resources_map imports webui_resources as "webui_resources.h" so our
-// overrides aren't included.
-#include "ui/resources/grit/webui_resources.h"
-
-#include "../gen/ui/resources/grit/webui_resources_map.cc"
