+++ components/brave_rewards/resources/page/brave_rewards_page.html
@@ -13,7 +13,7 @@
   <script src="chrome://resources/js/load_time_data_deprecated.js"></script>
   <script type="module" src="/strings.m.js"></script>
   <script src="/strings.js"></script>
-  <script type="module" src="chrome://resources/brave/br_elements/br_toolbar/br_toolbar.js"></script>
+  <script type="module" src="chrome://resources/cr_elements/cr_toolbar/cr_toolbar.js"></script>
   <script src="/brave_rewards_page.bundle.js"></script>
 </head>
 <body>
