+++ patches/tools-grit-preprocess_if_expr.py.patch
@@ -1,0 +1,28 @@
+diff --git a/tools/grit/preprocess_if_expr.py b/tools/grit/preprocess_if_expr.py
+index 35c1b20bc2c83..e7e239cb85457 100644
+--- a/tools/grit/preprocess_if_expr.py
++++ b/tools/grit/preprocess_if_expr.py
+@@ -82,6 +82,7 @@ def main(argv):
+   parser.add_argument('-E', '--environment')
+   parser.add_argument('-t', '--target')
+   parser.add_argument('--enable_removal_comments', action='store_true')
++  parser.add_argument('--overrides', nargs="*"); parser.add_argument('--override-in-folder')
+   args = parser.parse_args(argv)
+ 
+   in_folder = os.path.normpath(os.path.join(_CWD, args.in_folder))
+@@ -93,6 +94,7 @@ def main(argv):
+ 
+   for input_file in args.in_files:
+     in_path = os.path.join(in_folder, input_file)
++  def process_file(in_path, input_file):
+     content = ""
+     with open(in_path, encoding='utf-8') as f:
+       content = f.read()
+@@ -148,6 +150,6 @@ def main(argv):
+     json.dump(manifest_data, manifest_file)
+   return
+ 
+-
++from brave_chromium_utils import inline_chromium_src_override; inline_chromium_src_override(globals(), locals())
+ if __name__ == '__main__':
+   main(sys.argv[1:])
