+++ patches/ui-webui-resources-cr_elements-BUILD.gn.patch
@@ -1,12 +1,12 @@
 diff --git a/ui/webui/resources/cr_elements/BUILD.gn b/ui/webui/resources/cr_elements/BUILD.gn
-index c7103ae0d0a71d5f8c3371cfa85f1c3774721e1e..1c65e4e61175245c9254498c90b4a787a5c3fd0d 100644
+index c7103ae0d0a71..2f10eee24007e 100644
 --- a/ui/webui/resources/cr_elements/BUILD.gn
 +++ b/ui/webui/resources/cr_elements/BUILD.gn
 @@ -234,6 +234,7 @@ build_webui("build") {
      "//third_party/lit/v3_0:build_ts",
    ]
    if (!is_android && !is_ios) {
-+    ts_deps += ["//brave/ui/webui/resources:build_ts"]
++    import("//brave/ui/webui/resources/cr_elements/overrides.gni") ts_deps += ["//brave/ui/webui/resources:build_ts"] override_files = brave_cr_elements_override_files replace_files = brave_cr_elements_replace_files
      ts_deps += [ "//third_party/polymer/v3_0:library" ]
    }
    ts_definitions = [
