+++ script/override_utils.py
@@ -6,6 +6,7 @@
 import contextlib
 import inspect
 import types
+import sys
 
 from typing import Any
 
@@ -126,3 +127,31 @@ def override_scope_variable(scope,
             setattr(scope, name, original_value)
         else:
             delattr(scope, name)
+
+
+def call_function_get_frame(func, *args, **kwargs):
+    """
+    Calls the function *func* with the specified arguments and keyword
+    arguments and snatches its local frame before it actually executes.
+    Taken from: https://stackoverflow.com/a/52358426
+
+    This is useful if you need to override a function but use some of its
+    internal state.
+    """
+
+    frame = None
+    trace = sys.gettrace()
+
+    def snatch_locals(_frame, name, _):
+        nonlocal frame
+        if frame is None and name == 'call':
+            frame = _frame
+            sys.settrace(trace)
+        return trace
+
+    sys.settrace(snatch_locals)
+    try:
+        result = func(*args, **kwargs)
+    finally:
+        sys.settrace(trace)
+    return frame, result
