--- ui/webui/resources/br_elements/br_toolbar/br_toolbar_selection_overlay_overrides.css
@@ -1,3 +1,0 @@
-:host {
-  z-index: 3;
-}
