+++ ui/webui/resources/cr_elements/cr_toolbar/cr_toolbar.css
@@ -5,7 +5,7 @@
 
 /* #css_wrapper_metadata_start
  * #type=style-lit
- * #import=../br_shared_style.css.js
+ * #scheme=relative
  * #css_wrapper_metadata_end */
 
 :host {
