+++ ui/webui/resources/cr_elements/cr_toolbar/cr_toolbar_search_field.css
@@ -5,6 +5,7 @@
 
 /* #css_wrapper_metadata_start
  * #type=style-lit
+ * #scheme=relative
  * #css_wrapper_metadata_end */
 
 :host {
