+++ ui/webui/resources/cr_elements/cr_toolbar/cr_toolbar_selection_overlay.css
@@ -1,0 +1,15 @@
+/* Copyright (c) 2024 The Brave Authors. All rights reserved.
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this file,
+ * You can obtain one at https://mozilla.org/MPL/2.0/. */
+
+/* #css_wrapper_metadata_start
+ * #type=style-lit
+ * #import=./cr_toolbar_selection_overlay-chromium.css.js
+ * #scheme=relative
+ * #include=cr-toolbar-selection-overlay-chromium
+ * #css_wrapper_metadata_end */
+
+:host {
+  z-index: 3;
+}
