+++ ui/webui/resources/polymer_overriding.ts
@@ -6,7 +6,6 @@
 // @ts-nocheck TODO(petemill): Define types
 
 // Global overrides
-import './br_elements/br_toolbar/br_toolbar.js'
 import './br_elements/br_shared_style.css.js'
 
 import { html as polymerHtml, mixinBehaviors, Polymer, PolymerElement } from 'chrome://resources/polymer/v3_0/polymer/polymer_bundled.min.js';
